fix home page style
