+++ client/src/pages/Home/style.js
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const PageHead = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: center;
+`;
+
+export const HeadContainer = styled.div`
+  max-width: 425px;
+`;
+
+export const DropdownContainer = styled.div`
+  max-width: 300px;
+`;
